impls/add.rs step 2
